cleanup return types for APIs

+++ src/iso_alloc_interfaces.c
@@ -186,7 +186,7 @@ EXTERNAL_API void iso_alloc_unprotect_root() {
     _iso_alloc_unprotect_root();
 }
 
-EXTERNAL_API int64_t iso_alloc_detect_zone_leaks(iso_alloc_zone_handle *zone) {
+EXTERNAL_API uint64_t iso_alloc_detect_zone_leaks(iso_alloc_zone_handle *zone) {
     if(zone == NULL) {
         return 0;
     } else {
@@ -194,19 +194,19 @@ EXTERNAL_API int64_t iso_alloc_detect_zone_leaks(iso_alloc_zone_handle *zone) {
     }
 
     LOCK_ROOT_MUTEX();
-    int32_t r = _iso_alloc_zone_leak_detector(zone);
+    uint64_t r = _iso_alloc_zone_leak_detector(zone);
     UNLOCK_ROOT_MUTEX();
     return r;
 }
 
-EXTERNAL_API int64_t iso_alloc_detect_leaks() {
+EXTERNAL_API uint64_t iso_alloc_detect_leaks() {
     LOCK_ROOT_MUTEX();
-    int32_t r = _iso_alloc_detect_leaks();
+    uint64_t r = _iso_alloc_detect_leaks();
     UNLOCK_ROOT_MUTEX();
     return r;
 }
 
-EXTERNAL_API int64_t iso_alloc_zone_mem_usage(iso_alloc_zone_handle *zone) {
+EXTERNAL_API uint64_t iso_alloc_zone_mem_usage(iso_alloc_zone_handle *zone) {
     if(zone == NULL) {
         return 0;
     } else {
@@ -214,14 +214,14 @@ EXTERNAL_API int64_t iso_alloc_zone_mem_usage(iso_alloc_zone_handle *zone) {
     }
 
     LOCK_ROOT_MUTEX();
-    int32_t r = _iso_alloc_zone_mem_usage(zone);
+    uint64_t r = _iso_alloc_zone_mem_usage(zone);
     UNLOCK_ROOT_MUTEX();
     return r;
 }
 
-EXTERNAL_API int64_t iso_alloc_mem_usage() {
+EXTERNAL_API uint64_t iso_alloc_mem_usage() {
     LOCK_ROOT_MUTEX();
-    int32_t r = _iso_alloc_mem_usage();
+    uint64_t r = _iso_alloc_mem_usage();
     UNLOCK_ROOT_MUTEX();
     return r;
 }
